pyproject.toml
[project]
name = "agentify-example-aeo-bench"
version = "0.1.0"
description = "AEO-Bench: Answer Engine Optimization benchmark for evaluating documentation generation agents."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "dotenv>=0.9.9",
    "earthshaker",
    "langgraph>=0.2.0",
    "litellm>=1.0.0",
    "typer>=0.19.2",
    "uvicorn>=0.37.0",
]
